fix bugs, add method contactServiceImpl

+++ src/main/java/com/sgkhmjaes/jdias/service/impl/ContactServiceImpl.java
@@ -2,18 +2,21 @@ package com.sgkhmjaes.jdias.service.impl;
 
 import com.sgkhmjaes.jdias.service.ContactService;
 import com.sgkhmjaes.jdias.domain.Contact;
+import com.sgkhmjaes.jdias.domain.Person;
 import com.sgkhmjaes.jdias.repository.ContactRepository;
 import com.sgkhmjaes.jdias.repository.search.ContactSearchRepository;
+import com.sgkhmjaes.jdias.service.PersonService;
+import com.sgkhmjaes.jdias.service.UserService;
+import java.util.ArrayList;
+import java.util.Iterator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
 import java.util.List;
-import java.util.stream.Collectors;
-import java.util.stream.StreamSupport;
-
+import java.util.Set;
 import static org.elasticsearch.index.query.QueryBuilders.*;
+import org.hibernate.Hibernate;
 
 /**
  * Service Implementation for managing Contact.
@@ -25,12 +28,16 @@ public class ContactServiceImpl implements ContactService{
     private final Logger log = LoggerFactory.getLogger(ContactServiceImpl.class);
 
     private final ContactRepository contactRepository;
-
     private final ContactSearchRepository contactSearchRepository;
+    private final UserService userService;
+    private final PersonService personService;
 
-    public ContactServiceImpl(ContactRepository contactRepository, ContactSearchRepository contactSearchRepository) {
+    public ContactServiceImpl(ContactRepository contactRepository, ContactSearchRepository contactSearchRepository, 
+            PersonService personService, UserService userService) {
         this.contactRepository = contactRepository;
         this.contactSearchRepository = contactSearchRepository;
+        this.userService = userService;
+        this.personService = personService;
     }
 
     /**
@@ -42,6 +49,23 @@ public class ContactServiceImpl implements ContactService{
     @Override
     public Contact save(Contact contact) {
         log.debug("Request to save Contact : {}", contact);
+        
+        Person recipientContact = contact.getPerson();
+        Person currentPerson = userService.getCurrentPerson();
+        
+        contact.setOwnId(recipientContact.getId());
+        contact.setRecipient(recipientContact.getDiasporaId());
+        contact.setAuthor(currentPerson.getDiasporaId());
+        contact.setPerson(currentPerson);
+        
+        if (currentPerson.getContacts().contains(contact)) {
+            for (Contact currentPersonContact : currentPerson.getContacts()) 
+                if (currentPersonContact.equals(contact)) return currentPersonContact;
+        }
+        
+        currentPerson.addContacts(contact);
+        personService.save(currentPerson);
+        
         Contact result = contactRepository.save(contact);
         contactSearchRepository.save(result);
         return result;
@@ -56,7 +80,10 @@ public class ContactServiceImpl implements ContactService{
     @Transactional(readOnly = true)
     public List<Contact> findAll() {
         log.debug("Request to get all Contacts");
-        return contactRepository.findAll();
+        Person currentPerson = userService.getCurrentPerson();
+        Set<Contact> contacts = currentPerson.getContacts();
+        Hibernate.initialize(contacts);
+        return new ArrayList<>(contacts);
     }
 
     /**
@@ -69,7 +96,11 @@ public class ContactServiceImpl implements ContactService{
     @Transactional(readOnly = true)
     public Contact findOne(Long id) {
         log.debug("Request to get Contact : {}", id);
-        return contactRepository.findOne(id);
+        Person currentPerson = userService.getCurrentPerson();
+        Set<Contact> contacts = currentPerson.getContacts();
+        Contact findContact = contactRepository.findOne(id);
+        if (contacts.contains(findContact)) return findContact;
+        else return new Contact();
     }
 
     /**
@@ -80,8 +111,16 @@ public class ContactServiceImpl implements ContactService{
     @Override
     public void delete(Long id) {
         log.debug("Request to delete Contact : {}", id);
-        contactRepository.delete(id);
-        contactSearchRepository.delete(id);
+        Person currentPerson = userService.getCurrentPerson();
+        Set<Contact> contacts = currentPerson.getContacts();
+        Contact findContact = contactRepository.findOne(id);
+        if (contacts.contains(findContact)) {
+            contacts.remove(findContact);
+            currentPerson.setContacts(contacts);
+            personService.save(currentPerson);
+            contactRepository.delete(id);
+            contactSearchRepository.delete(id);
+        }
     }
 
     /**
@@ -94,8 +133,24 @@ public class ContactServiceImpl implements ContactService{
     @Transactional(readOnly = true)
     public List<Contact> search(String query) {
         log.debug("Request to search Contacts for query {}", query);
+        List <Contact> findContacts = new ArrayList<>();
+        Person currentPerson = userService.getCurrentPerson();
+        Set<Contact> contacts = currentPerson.getContacts();
+        
+        Iterable<Contact> search = contactSearchRepository.search(queryStringQuery(query));
+        Iterator<Contact> iterator = search.iterator();
+        while (iterator.hasNext()) {
+            Contact findContact = iterator.next();
+            if (contacts.contains(findContact)) findContacts.add(findContact);
+        }
+        return findContacts;
+        /*
         return StreamSupport
             .stream(contactSearchRepository.search(queryStringQuery(query)).spliterator(), false)
-            .collect(Collectors.toList());
+            .collect(Collectors.toList());*/
+    }
+    
+    public Person getRecipientPerson (Contact contact) {
+        return personService.findOne(contact.getOwnId());
     }
 }
